refactor(sdk): unify DefenderIdentity into defender service
Move the health issues check from defenderidentity to defender service,
following the S3/S3Control architectural pattern where multiple service
classes coexist in a single service directory.

- Add DefenderIdentity class and HealthIssue model to defender_service.py
- Create defender_identity_client.py singleton
- Rename check to defender_identity_health_issues_no_open
- Update compliance framework reference in iso27001_2022_m365.json
- Delete old defenderidentity service directory
- Update all test imports and paths
