+++ prowler/providers/m365/services/defender/defender_identity_health_issues_no_open/defender_identity_health_issues_no_open.metadata.json
@@ -1,16 +1,16 @@
 {
   "Provider": "m365",
-  "CheckID": "defenderidentity_health_issues_no_open",
-  "CheckTitle": "Defender for Identity monitors hybrid infrastructure without unresolved health issues",
+  "CheckID": "defender_identity_health_issues_no_open",
+  "CheckTitle": "Ensure Defender for Identity monitors hybrid infrastructure without unresolved health issues",
   "CheckType": [],
-  "ServiceName": "defenderidentity",
-  "SubServiceName": "",
+  "ServiceName": "defender",
+  "SubServiceName": "identity",
   "ResourceIdTemplate": "",
   "Severity": "medium",
   "ResourceType": "Defender for Identity Health Issue",
   "ResourceGroup": "security",
-  "Description": "Microsoft Defender for Identity (MDI) monitors your hybrid identity infrastructure and detects advanced threats targeting Active Directory.\n\nHealth issues indicate problems with the MDI configuration or sensors that may affect the ability to detect identity-based attacks.",
-  "Risk": "Unresolved health issues in Defender for Identity can lead to:\n- **Gaps in threat detection** - Misconfigured sensors may fail to monitor domain controllers\n- **Delayed attack identification** - Identity-based attacks like Pass-the-Hash or Golden Ticket may go undetected\n- **Compromised hybrid infrastructure** - Attackers could exploit blind spots in identity monitoring",
+  "Description": "Microsoft Defender for Identity (MDI) monitors your hybrid identity infrastructure and detects advanced threats targeting Active Directory. Health issues indicate problems with the MDI configuration or sensors that may affect the ability to detect identity-based attacks.",
+  "Risk": "Unresolved health issues in Defender for Identity can lead to gaps in threat detection where misconfigured sensors may fail to monitor domain controllers, delayed attack identification where identity-based attacks like Pass-the-Hash or Golden Ticket may go undetected, and compromised hybrid infrastructure where attackers could exploit blind spots in identity monitoring.",
   "RelatedUrl": "",
   "AdditionalURLs": [
     "https://learn.microsoft.com/en-us/defender-for-identity/health-alerts",
@@ -25,7 +25,7 @@
     },
     "Recommendation": {
       "Text": "Regularly monitor and resolve Defender for Identity health issues to maintain comprehensive visibility into identity-based threats across your hybrid infrastructure.",
-      "Url": "https://hub.prowler.com/check/defenderidentity_health_issues_no_open"
+      "Url": "https://learn.microsoft.com/en-us/defender-for-identity/health-alerts"
     }
   },
   "Categories": [
